requirements.txt
Django>=5.0
djangorestframework>=3.15
django-filter>=24.2
djangorestframework-simplejwt>=5.3
python-dotenv>=1.0

